LoFTR-swa/requirements.txt
opencv-python-headless==4.5.5.62
albumentations==0.5.1
ray>=1.0.1
einops==0.3.0
kornia==0.4.1
loguru==0.5.3
yacs>=0.1.8
tqdm
autopep8
pylint
ipython
jupyterlab
matplotlib
h5py==3.7.0
pytorch-lightning==1.3.5
torchmetrics==0.6.0
joblib>=1.0.1
